stable release GH action fixes
Assuming v0.1 is the stable version.
Now if a new code or a backport is pushed ot the v0.1 branch then the
stable image should also be updated. This was not getting updated
because the stable image push was done even before the v0.1 image was
uploaded (since the actions were operating in parallel). We now have
added the stable release update as a different job in the same action
yaml thus ensuring that it executes serially.

Signed-off-by: Rahul Jadhav <nyrahul@gmail.com>